Fix DMA callback references for memory buffers
For the BDMA the order of the half completed interrupt calls seems in the wrong order.

If I am at a HALF interrupt I have also to call the corresponding interrupt of the SAME buffer (because it is obviously not completed yet). Compare with the Half Transfer Complete Interrupt management of the regular DMA.

Just stumbled upon this without further research. You might want to check your entire code base for the correct call assignment of Xfer*CpltCallback depending on DMA_SxCR_CT.